// frontend/src/api/generated.ts
export interface components {
  schemas: {
// ...
    ImportCommitResponse: {
      job: components["schemas"]["ImportJobResponse"];
      /** Createdcustomercount */
      createdCustomerCount: number;
      /** Createdinteractioncount */
      createdInteractionCount: number;
    };
// ...
    ImportPreviewResponse: {
      job: components["schemas"]["ImportJobResponse"];
      /** Errors */
      errors: components["schemas"]["ImportErrorResponse"][];
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}